--- src/components/common/neighbor/List.js
@@ -1,16 +1,0 @@
-import styled from 'styled-components';
-import { Underline } from './Header';
-
-export const LongUnderline = styled(Underline)`
-  width: 379px;
-`;
-
-export const NeighborList = styled.div`
-  display: flex;
-  align-items: center;
-  height: 47px;
-`;
-
-export const Button = styled.button`
-  font-size: 18px;
-`;
